requirements.txt
streamlit
presidio-analyzer
presidio-anonymizer
spacy
google-generativeai
pyyaml
